+++ components/task-edit.tsx
@@ -10,7 +10,7 @@ import {
   searchParams,
   session,
   setRefresh,
-} from "./session";
+} from "./session-provider";
 import * as z from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
